Fixed Update
Previously before Rendering each UpdateComponent was updated before Rendering.

This led to Uniforms being overwritten.
e.g. You have 2 Entities with TransformComponents the first one will not work because the u_ObjectTransform Uniform was overwritten by the TransformComponent of the 2. Entity.

Changed it to Updating each Component of an Entity and rendering the Entity after the Updates.